requirements.txt
bolt-ta
conttest
coverage
m2r2
pytest
pytest-cov
sphinx
sphinx-rtd-theme
